requirements.txt
pandas
numpy
requests
fastapi
uvicorn
sqlalchemy
pydantic
alembic
pytest
pytest-cov
black
pymongo[srv]==3.12
psycopg2-binary
faker==20.1.0
python-dotenv==1.0.0
pymongo==4.6.1
python-dotenv==1.0.0
dbt-postgres
